Bug 18267 populate_order_with_prices should be exported
Although routine is used in acqui scripts, access to it
is by backdoor methods. Allow it to be exported in a standard way
to allow more maintainable code.

Effectively this patch merely documents that this routine
is part of C4/Acquisition's interface and allows users to import
it in the standard way.