Bug 18267 Populate new price fields in EDI Processing
New price fields have been added to aqorders with the result
that sites using EDI found prices appearing as zero in
displays as the prices were not duplicated to the
nrew fields as in manual acquisitions.

Invoicing now records values using the standard
populate_order_with_prices routine from C4::Acquisitions

However when creating an order from an edifact quote the
routine makes invalid assumptions about what data is available.
In a quote prices are always tax inclusive, and no information
is supplied on the tax rate and amount, these appear in the invoice
message on receipting. Koha::EDI therefore sets the relevant
fields directly in order creation.

To clarify the field used, we now call the price_info method rather than
the older price method.

When calculating discount, we now check if a discounted price
has been supplied. (In practice we dont see this but we are
prepared if it occurs )