Bug 18267 Remove obsolete Edifact::Line price method
The old price method duplicated the function of price_info as this
was the only type of price field in quotes.
The more specific field is now used and the old method is
no longer called. Remove the obsolete code

Add comment documenting the allowed values in the Edifact
PRI segment