Bug 18267 - Record tax rate on receipt from EDI Invoice
As part of the newly introduced price fields there
is also a tax_rate_on_receiving that needs setting.

This records the main tax rate on receipting. Edifact has the
potential to be more complex than Koha allows i.e. the
line amount may be made of different amounts (incorporating
servicing, etc) at different tax rates. This concentrates
on the main tax rate.

The tax rate will be receiced as a percentage value e.g. 20.00
need to convert to a decimal fraction 0.20 as assumed in Koha
calculations