(pm) Swap Location and Office labels
Request from pm:
On routing lists, the ‘Library’ field and the ‘Address’ field from the patron record
display as ‘Location’ and ‘Office’ respectively. This might prove confusing for
our post room staff – could we have the column labels switched round, so that ‘Office’
would be the Library, and ‘Location’ would be the Address?
We could then store lawyers’ desk/pigeonhole numbers in the Address field