Bug 21988: Add support for SMSSendFrom syspref
This patch adds a SMSSendFrom syspref which allows a "from" number to be
provided to SMS drivers that support it. It is passed to the driver as a
"_from" parameter, in line with the formatting of the _username and
_password parameters.

Testing:

Since this work depends on an SMS driver that supports it, it is
virtually impossible to test. I have added the mocking of the
SMSSendFrom preference to t/SMS.t, but can't really take it beyond
that. Other than the adding of the syspref, a very tiny modification is
made to C4/SMS.pm, but I don't anticipate this having an adverse effect.
SMS drivers that do not support the _from paramter will simple ignore
it.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>