Bug 27378: Add new 'modaljs' syspref type
This commit adds a new 'modaljs' syspref type. This is similar to the existing
'modalselect' syspref type but instead of using a file to create a list of
selectable options which are displayed in the modal, it assumes that anything
can happen within the modal driven by JS, and whatever happens is going to
return a string, that will be saved as the syspref value. In our case, the
string will be all the details of the “chunks” concatenated into a string

Signed-ff-by: Barry Cannon <bc@interleaf.ie>